--- test/index.spec.ts
@@ -1,579 +1,0 @@
-import { expect } from 'chai';
-import { Sparsely, ConfigOption, ParsedOption } from '../lib';
-import testOptions from './test-options';
-
-describe('Sparse (without test options)', () => {
-
-  const parser = new Sparsely();
-
-  describe('options', () => {
-    it ('should return an empty array when no options have been added', () => {
-      const options = parser.options;
-      expect(options).to.have.lengthOf(0);
-    });
-  });
-
-  describe('parsedArgs', () => {
-    it('should return an empty array when exec has not been called', () => {
-      const parsedArgs = parser.parsedArgs;
-      expect(parsedArgs).to.have.lengthOf(0);
-    });
-  });
-
-  describe('parsedOptions', () => {
-    it('should return an empty array when exec has not been called', () => {
-      const parsedOptions = parser.parsedOptions;
-      expect(parsedOptions).to.have.lengthOf(0);
-    });
-  });
-
-  describe('errors', () => {
-    it('should return an empty array when exec has not been called', () => {
-      const errors = parser.errors;
-      expect(errors).to.have.lengthOf(0);
-    });
-  });
-});
-
-describe('Sparse (with test options)', () => {
-  // This suite simply tests that options
-  // are added correctly via the addOption
-  // method:
-  describe('options', () => {
-    const parser = new Sparsely();
-
-    testOptions.forEach((option: ConfigOption )=> {
-      parser.addOption(option);
-    });
-
-    it(`should return an array of length ${testOptions.length}`, () => {
-      const options = parser.options;
-      expect(options).to.have.lengthOf(options.length);
-    });
-  });
-
-  // This suite tests the scenario where only arguments
-  // have been passed to the program using Sparse:
-  describe('argv = ["arg1", "arg2", "arg3"]', () => {
-    const parser = new Sparsely();
-
-    testOptions.forEach((option: ConfigOption) => {
-      parser.addOption(option);
-    });
-
-    const argv = [ 'arg1', 'arg2', 'arg3' ];
-    parser.exec(argv);
-
-    describe('errors', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const errors = parser.errors;
-        expect(errors).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedArgs', () => {
-      const expected = argv.length;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedArgs = parser.parsedArgs;
-        expect(parsedArgs).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedOptions', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedOptions = parser.parsedOptions;
-        expect(parsedOptions).to.have.lengthOf(expected);
-      });
-    });
-  });
-
-  // This suite tests to ensure that
-  // a single shorthand option that 
-  // does not accept arguments is 
-  // parsed correctly in verbose form:
-  describe('argv = ["--option-A"]', () => {
-    const parser = new Sparsely();
-
-    testOptions.forEach((option: ConfigOption) => {
-      parser.addOption(option);
-    });
-
-    const argv = [ '-A' ];
-    parser.exec(argv);
-
-    describe('errors', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const errors = parser.errors;
-        expect(errors).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedArgs', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedArgs = parser.parsedArgs;
-        expect(parsedArgs).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedOptions', () => {
-      const expected = 1;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedOptions = parser.parsedOptions;
-        expect(parsedOptions).to.have.lengthOf(expected);
-      });
-    });
-  });
-
-  // This suite tests to ensure that
-  // a single shorthand option that 
-  // does not accept arguments is 
-  // parsed correctly in shorthand form:
-  describe('argv = ["-A"]', () => {
-    const parser = new Sparsely();
-
-    testOptions.forEach((option: ConfigOption) => {
-      parser.addOption(option);
-    });
-
-    const argv = [ '-A' ];
-    parser.exec(argv);
-
-    describe('errors', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const errors = parser.errors;
-        expect(errors).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedArgs', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedArgs = parser.parsedArgs;
-        expect(parsedArgs).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedOptions', () => {
-      const expected = 1;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedOptions = parser.parsedOptions;
-        expect(parsedOptions).to.have.lengthOf(expected);
-      });
-    });
-  });
-
-  // This suite tests to ensure that
-  // multiple serparate verbose options
-  // are parsed correctly:
-  describe('argv = ["--option-A", "--option-F"]', () => {
-    const parser = new Sparsely();
-    testOptions.forEach((option: ConfigOption) => {
-      parser.addOption(option);
-    });
-
-    const argv = [ '--option-A', '--option-F' ];
-    parser.exec(argv);
-
-    describe('errors', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const errors = parser.errors;
-        expect(errors).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedArgs', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedArgs = parser.parsedArgs;
-        expect(parsedArgs).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedOptions', () => {
-      const expected = 2;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedOptions = parser.parsedOptions;
-        expect(parsedOptions).to.have.lengthOf(expected);
-      });
-    });
-  });
-
-  // This suite tests to ensure that
-  // multiple serparate shorthand options
-  // are parsed correctly:
-  describe('argv = ["-A", "-F"]', () => {
-    const parser = new Sparsely();
-    testOptions.forEach((option: ConfigOption) => {
-      parser.addOption(option);
-    });
-
-    const argv = [ '-A', '-F' ];
-    parser.exec(argv);
-
-    describe('errors', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const errors = parser.errors;
-        expect(errors).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedArgs', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedArgs = parser.parsedArgs;
-        expect(parsedArgs).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedOptions', () => {
-      const expected = 2;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedOptions = parser.parsedOptions;
-        expect(parsedOptions).to.have.lengthOf(expected);
-      });
-    });
-  });
-
-  // This suite tests to ensure that
-  // a group of valid options is parsed
-  // correctly:
-  describe('argv = ["-AF"]', () => {
-    const parser = new Sparsely();
-
-    testOptions.forEach((option: ConfigOption) => {
-      parser.addOption(option);
-    });
-
-    const argv = [ '-AF' ];
-    parser.exec(argv);
-
-    describe('errors', () => {
-      const expected = 2;
-      it(`should return an array of length ${expected}`, () => {
-        const errors = parser.errors;
-        expect(errors).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedArgs', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedArgs = parser.parsedArgs;
-        expect(parsedArgs).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedOptions', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedOptions = parser.parsedOptions;
-        expect(parsedOptions).to.have.lengthOf(expected);
-      });
-    });
-  });
-
-  // This suite tests to ensure that
-  // a verbose option that accepts
-  // an argument is parsed correctly:
-  describe('argv = ["--option-B", "arg1"]', () => {
-    const parser = new Sparsely();
-
-    testOptions.forEach((option: ConfigOption) => {
-      parser.addOption(option);
-    });
-
-    const argv = [ '--option-B', 'arg1' ];
-    parser.exec(argv);
-
-    describe('errors', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const errors = parser.errors;
-        expect(errors).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedArgs', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedArgs = parser.parsedArgs;
-        expect(parsedArgs).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedOptions', () => {
-      const expected = 1;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedOptions = parser.parsedOptions;
-        expect(parsedOptions).to.have.lengthOf(expected);
-      });
-    });
-  });
-
-  // This suite tests to ensure that
-  // a shorthand option that accepts
-  // an argument is parsed correctly:
-  describe('argv = ["-B", "arg1"]', () => {
-    const parser = new Sparsely();
-
-    testOptions.forEach((option: ConfigOption) => {
-      parser.addOption(option);
-    });
-
-    const argv = [ '-B', 'arg1' ];
-    parser.exec(argv);
-
-    describe('errors', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const errors = parser.errors;
-        expect(errors).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedArgs', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedArgs = parser.parsedArgs;
-        expect(parsedArgs).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedOptions', () => {
-      const expected = 1;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedOptions = parser.parsedOptions;
-        expect(parsedOptions).to.have.lengthOf(expected);
-      });
-    });
-  });
-
-  // This suite tests to ensure that
-  // a verbose option that accepts
-  // an argument is parsed correctly
-  // when in key/value form:
-  describe('argv = ["--option-C=arg"]', () => {
-    const parser = new Sparsely();
-
-    testOptions.forEach((option: ConfigOption) => {
-      parser.addOption(option);
-    });
-
-    const argv = [ '--option-C=arg' ];
-    parser.exec(argv);
-
-    describe('errors', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const errors = parser.errors;
-        expect(errors).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedArgs', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedArgs = parser.parsedArgs;
-        expect(parsedArgs).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedOptions', () => {
-      const expected = 1;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedOptions = parser.parsedOptions;
-        expect(parsedOptions).to.have.lengthOf(expected);
-      });
-    });
-  });
-  
-  // This suite tests to ensure that
-  // all types of arguments are parsed
-  // correctly:
-  describe('argv = ["-AF", "--option-B", "option-B-arg", "--option-C=arg", "program-arg-1", "program-arg-2", "-G", "option-G-arg-1", "option-G-arg-2", "-E", "-D"]', () => {
-    const parser = new Sparsely();
-
-    testOptions.forEach((option: ConfigOption) => {
-      parser.addOption(option);
-    });
-
-    const argv = [ 
-      '-ED', 
-      '--option-B', 
-      'option-B-arg',
-      '--option-C=arg',
-      'program-arg-1',
-      'program-arg-2',
-      '-G',
-      'option-G-arg-1',
-      '-A',
-      '-F'
-    ];
-    parser.exec(argv);
-
-    describe('errors', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, ()=> {
-        const errors = parser.errors;
-        expect(errors).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedArgs', () => {
-      const expected = 2;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedArgs = parser.parsedArgs;
-        expect(parsedArgs).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedOptions', () => {
-      const expected = 7;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedOptions = parser.parsedOptions;
-        expect(parsedOptions).to.have.lengthOf(expected);
-      });
-    });
-  });
-
-  // This suite tests to ensure that
-  // a verbose option is parsed correctly
-  // when insufficient arguments are 
-  // provided to it (appropriate error
-  // messages are generated):  
-  describe('argv = ["--option-B"]', () => {
-    const parser = new Sparsely();
-
-    testOptions.forEach((option: ConfigOption) => {
-      parser.addOption(option);
-    });
-
-    const argv = [ '--option-B' ];
-    parser.exec(argv);
-
-    describe('errors', () => {
-      const expected = 1;
-      it(`should return an array of length ${expected}`, () => {
-        const errors = parser.errors;
-        expect(errors).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedArgs', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedArgs = parser.parsedArgs;
-        expect(parsedArgs).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedOptions', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedOptions = parser.parsedOptions;
-        expect(parsedOptions).to.have.lengthOf(expected);
-      });
-    });
-  });
-
-  // This suite tests to ensure that
-  // a shorthand option is parsed correctly
-  // when insufficient arguments are provided
-  // to it (appropriate error messages are
-  // generated):
-  describe('argv = ["-B"]', () => {
-    const parser = new Sparsely();
-
-    testOptions.forEach((option: ConfigOption) => {
-      parser.addOption(option);
-    });
-
-    const argv = [ '-B' ];
-    parser.exec(argv);
-
-    describe('errors', () => {
-      const expected = 1;
-      it(`should return an array of length ${expected}`, () => {
-        const errors = parser.errors;
-        expect(errors).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedArgs', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedArgs = parser.parsedArgs;
-        expect(parsedArgs).to.have.lengthOf(expected);
-      });
-    });
-
-    describe('parsedOptions', () => {
-      const expected = 0;
-      it(`should return an array of length ${expected}`, () => {
-        const parsedOptions = parser.parsedOptions;
-        expect(parsedOptions).to.have.lengthOf(expected);
-      });
-    });
-  });
-
-  // This suite tests to ensure that
-  // the isParsedOption method indicates
-  // whether an option has been parsed
-  // or not:
-  describe('argv = ["-A", "-B"]', () => {
-    const parser = new Sparsely();
-
-    testOptions.forEach((option: ConfigOption) => {
-      parser.addOption(option);
-    });
-
-    const argv = [ '-A', '-B' ];
-    parser.exec(argv);
-
-    describe('isParsedOption', () => {
-      it('should return true with an argument of "option-A"', () => {
-        const result = parser.isParsedOption('option-A');
-        expect(result).to.be.true;
-      });
-
-      it('should return false with an argument of "option-B"', () => {
-        const result = parser.isParsedOption('option-B');
-        expect(result).to.be.false;
-      });
-    });
-  });
-
-  // This suite tests to ensure that
-  // the isParsedOption method indicates
-  // whether an option has been parsed
-  // or not:
-  describe('argv = ["arg1"]', () => {
-    const parser = new Sparsely();
-
-    testOptions.forEach((option: ConfigOption) => {
-      parser.addOption(option);
-    });
-
-    const argv = [ 'arg1' ];
-    parser.exec(argv);
-
-    describe('isParsedArg', () => {
-      it('should return true with an argument of "arg1"', () => {
-        const result = parser.isParsedArg('arg1');
-        expect(result).to.be.true;
-      });
-
-      it('should return false with an argument of "arg2"', () => {
-        const result = parser.isParsedArg('arg2');
-        expect(result).to.be.false;
-      });
-    });
-  });
-});
